fix(test): CodeRabbit 再指摘対応と Team Leader Companion e2e の part 完了時レビュー反映
- agent-usecases.test: onActivity の activity 引数を ProviderActivityCallback 型で転送
- test-helpers: findWorkflowStep / findAgentWorkflowStep を共有ヘルパーへ集約し、workflow-call-schema / workflowLoader から import
- companion-review.e2e: Team Leader ケースの mock シナリオに part 完了時レビュー（initial / fix）の応答を追加し、security-reviewer の呼び出し回数を 4 に更新（Team scope の mailbox が書かれず ENOENT になっていた e2e-mock の失敗を修正）
